✅ 🐧🧹 🖳🗑️

+++ misc/Github/Runners/Self-Hosted/aarch64-ubuntu.dockerfile
@@ -151,7 +151,11 @@ EOS
 WORKDIR /runner
 RUN <<EOS
  #Setup GH Runner (arm64)
-  eget "https://github.com/actions/runner" --asset "linux" --asset "arm64" --asset "tar.gz" --to "./runner.tar.gz" --download-only 
+  #eget "https://github.com/actions/runner" --asset "linux" --asset "arm64" --asset "tar.gz" --to "./runner.tar.gz" --download-only
+  wget --quiet --show-progress "https://pub.ajam.dev/utils/gh-runner-aarch64/runner.tar.gz" -O "./runner.tar.gz"
+  if [ ! -f "./runner.tar.gz" ]; then
+     exit 1
+  fi
  #Untar
   mkdir -p "/runner-init"
   tar xzf "./runner.tar.gz" -C "/runner-init" && rm "./runner.tar.gz"
